Move FileUtils to lib-communicationTools to be used by client and services
